tensorflow2_mnist/requirements.txt
tensorflow==2.1.0
Pillow==7.0.0
